@@ -39,16 +39,7 @@ import { ThemeSetterModule } from './theme-setter/theme-setter.module';
     BrowserAnimationsModule,
     StoreModule.forRoot(reducers, {initialState}),
     // TODO consider putting this in environment
-    AngularFireModule.initializeApp({
-      apiKey: 'AIzaSyCKDvqohDwPW-Up1ZgQzASqUWDX8QfCA4s',
-      authDomain: 'crossclip.firebaseapp.com',
-      databaseURL: 'https://crossclip.firebaseio.com',
-      projectId: 'crossclip',
-      storageBucket: 'crossclip.appspot.com',
-      messagingSenderId: '447960949346',
-      appId: '1:447960949346:web:6a1575386630710539673b',
-      measurementId: 'G-4VZZLXSVKJ'
-    }),
+    AngularFireModule.initializeApp(environment.firebaseConfig),
     AngularFirestoreModule,
     AngularFireAuthGuardModule,
     StoreDevtoolsModule.instrument({maxAge: 25, logOnly: environment.production}),
