@@ -1,14 +1,19 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ProfileButtonPopupComponent } from './profile-button-popup.component';
+import {PROFILE_BUTTON_CALLBACKS, PROFILE_BUTTON_DISPLAY_NAME} from '../profile-button.component';
 
 describe('ProfileButtonPopupComponent', () => {
   let component: ProfileButtonPopupComponent;
   let fixture: ComponentFixture<ProfileButtonPopupComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ProfileButtonPopupComponent ]
+      declarations: [ ProfileButtonPopupComponent ],
+      providers: [
+        { provide: PROFILE_BUTTON_DISPLAY_NAME, useValue: 'test user' },
+        { provide: PROFILE_BUTTON_CALLBACKS, useValue: { onProfile: () => {}, onLogout: () => {} } },
+      ]
     })
     .compileComponents();
   }));
