@@ -1,13 +1,13 @@
-import { reducer, initialState } from './app-state.reducer';
+import { appStateReducer, appStateInitialState } from './app-state.reducer';
 
 describe('AppState Reducer', () => {
   describe('an unknown action', () => {
     it('should return the previous state', () => {
       const action = {} as any;
 
-      const result = reducer(initialState, action);
+      const result = appStateReducer(appStateInitialState, action);
 
-      expect(result).toBe(initialState);
+      expect(result).toBe(appStateInitialState);
     });
   });
 });
