@@ -1,13 +1,13 @@
-import { reducer, initialState } from './clip.reducer';
+import { clipsReducer, clipsInitialState } from './clip.reducer';
 
 describe('Clip Reducer', () => {
   describe('unknown action', () => {
     it('should return the previous state', () => {
       const action = {} as any;
 
-      const result = reducer(initialState, action);
+      const result = clipsReducer(clipsInitialState, action);
 
-      expect(result).toBe(initialState);
+      expect(result).toBe(clipsInitialState);
     });
   });
 });
