@@ -1,14 +1,22 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ThemeSetterComponent } from './theme-setter.component';
+import {MockStore, provideMockStore} from '@ngrx/store/testing';
+import {initialState} from '../state/state';
 
 describe('ThemeSetterComponent', () => {
   let component: ThemeSetterComponent;
   let fixture: ComponentFixture<ThemeSetterComponent>;
+  let mockStore: MockStore;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ThemeSetterComponent ]
+      declarations: [ ThemeSetterComponent ],
+      providers: [
+        provideMockStore({
+          initialState,
+        }),
+      ]
     })
     .compileComponents();
   }));
@@ -17,6 +25,8 @@ describe('ThemeSetterComponent', () => {
     fixture = TestBed.createComponent(ThemeSetterComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
+
+    mockStore = TestBed.inject(MockStore);
   });
 
   it('should create', () => {
