@@ -1,3 +1,13 @@
 export const environment = {
-  production: true
+  production: true,
+  firebaseConfig: {
+    apiKey: 'AIzaSyCKDvqohDwPW-Up1ZgQzASqUWDX8QfCA4s',
+    authDomain: 'crossclip.firebaseapp.com',
+    databaseURL: 'https://crossclip.firebaseio.com',
+    projectId: 'crossclip',
+    storageBucket: 'crossclip.appspot.com',
+    messagingSenderId: '447960949346',
+    appId: '1:447960949346:web:6a1575386630710539673b',
+    measurementId: 'G-4VZZLXSVKJ'
+  },
 };
