@@ -3,7 +3,17 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebaseConfig: {
+    apiKey: 'AIzaSyCKDvqohDwPW-Up1ZgQzASqUWDX8QfCA4s',
+    authDomain: 'crossclip.firebaseapp.com',
+    databaseURL: 'https://crossclip.firebaseio.com',
+    projectId: 'crossclip',
+    storageBucket: 'crossclip.appspot.com',
+    messagingSenderId: '447960949346',
+    appId: '1:447960949346:web:6a1575386630710539673b',
+    measurementId: 'G-4VZZLXSVKJ'
+  },
 };
 
 /*
